refactor: move utils out of drivers mod
